Don't skip propterty_notify for focused window
When receiving a property_notify event, ensure we update the focused
window(s) as well, and don't bail early.  If we skip the focused window then
urgency hint properties on focused windows on an unfocused monitor are
skipped.

This currently happens when there's a single window on an unfocused monitor
which then becomes focused by switching monitors, and the event is never
received.