Make the poll logic work on Linux
On Darwin we need to set the `events` field to `POLLHUP`, otherwise we
won't receive `POLLHUP` events (this doesn't make sense: here's an
excerpt from the *poll* manual regarding `POLLHUP`: *This flag is output
only, and ignored if present in the input _events_ bitmask*).

On Linux, we get a `POLLERR` event instead of the `POLLHUP` event. And
we could set the `events` field to 0, but then it wouldn't work on silly
systems that requires `POLLHUP`.