Don't honor pointer_follows_focus in apply_layout
The assumption here was that, when the focused node's geometry changes,
pointer_follows_focus needs to be taken care of. But we probably don't
want to do that when the pointer is still inside the node despite the
change of geometry. Of course this would also need to be taken care of
in {move,resize}_client, etc.

We chose the simpler approach here: only honor pointer_follows_focus in
focus_node.

Fixes #884.