Insert receptacle to replace stickies
Removing a sticky window from the desktop tree causes the entire layout
to go out the window, which makes sticky essentially useless for tiled
windows. Instead I suggest we insert a receptacle bound to the node when
moving a sticky window around. When moving that sticky window back
we reinsert it into that receptacle. This way we get to keep the layout
we set up, while still moving all the sticky windows around.

When removing the windows (or just the sticky flag) we now have to
remember to remove the placeholders from all desktops on that monitor.
There's probably many more cases where we have to remove the
placeholders. I'd like feedback on that.