Allow negated window modifiers to match non-wins
For example, `.!floating` used to only match nodes that hold a window
that is not in floating state; now, `.!floating` will match any node
that is not a node that holds a window that's in floating state.

The old behavior was counterintuitive, since e.g. `.tiled` and `.!tiled`
were not complementaries (they were complementaries with respect to the
windows, not the nodes) while the other standard node modifiers were.

Closes #1232.