Make negated window modifiers also match non-window nodes.
For example, `.!floating` used to only match nodes that hold a window
that is not in floating state; now, `.!floating` will match any node
that is not a node that holds a window that is in floating state.

The old behaviour can still be used by simply using the negated window
modifier with `.window`; e.g.:
* `.!floating` -> `.!floating.window`
* `.!same_class` -> `.!same_class.window`
* `.!hidden.!sticky` -> `.!hidden.!sticky.window`

In my opinion, the old behaviour was counterintuitive since, e.g.,
`.tiled` and `.!tiled` were not complementary (they were complementary
with respect to the windows, not the nodes) while the other standard
node modifiers were.