build: Add optional includes for BSD, and explicit rules for bspwm and bspc
Building BSPWM with `make` fails on OpenBSD 7.0 and FreeBSD 13.0. Errors include failure to find required headers and libraries, and failure to find make rules for `bspwm` and `bspc`. The issues regarding missing headers can be attributed to the differences in where xcb is installed on these systems, and the issue regarding `bspwm` and `bspc` can be attributed to the differences in implicit rules between the GNU and BSD versions of `make`.