Always return 1 when execvp() fails
bspwm uses execvp() to replace itself with a new instance in response to
"wm -r", and exits with the return value of execvp() if it fails.

execvp() only returns when it fails; and, when it returns, it always
returns -1 which is not a valid exit code (it is treated as 255: &0xff).

The return value of execvp() is not really meaningful; let's just exit
with `1' if execvp() fails.

Closes #1393.