_NET_WM_WINDOW_TYPE_UTILITY windows should float
According to:
https://specifications.freedesktop.org/wm-spec/1.5/ar01s05.html

>_NET_WM_WINDOW_TYPE_UTILITY indicates a small persistent utility
window, such as a palette or toolbox. It is distinct from type TOOLBAR
because it does not correspond to a toolbar torn off from the main
application. It's distinct from type DIALOG because it isn't a transient
dialog, the user will probably keep it open while they're working.
Windows of this type may set the WM_TRANSIENT_FOR hint indicating the
main application window.

This seems to indicate that this windows should start floating by
default, as one would expect of either a "palette" or a "toolbox".

Specifically, this has implications for Picture-In-Picture windows in
chromium.