Handle SIGCHLD instead of ignoring it
On Linux, if SIGCHLD is ignored, then it is still ignored after execve,
and ignoring SIGCHLD caused spawned processes to not be able to see
child statuses with waitpid.

This fixes a problem introduced with commit 6082d8b6, which I had only
tested on OpenBSD, where execve resets the ignored status of SIGCHLD.

Fixes #1533.
Closes #1534.